storage_types: extract RegionInfo, SeekRegionCallback
