+++ components/storage_types/Cargo.toml
@@ -7,7 +7,9 @@ publish = false
 [dependencies]
 byteorder = "1.2"
 hex = "0.3"
+kvproto = { git = "https://github.com/pingcap/kvproto.git" }
 quick-error = "1.2.2"
+raft = "0.6.0-alpha"
 rand = "0.6.5"
 slog = { version = "2.3", features = ["max_level_trace", "release_max_level_debug"] }
 slog-global = { version = "0.1", git = "https://github.com/breeswish/slog-global.git", rev = "0e23a5baff302a9d7bccd85f8f31e43339c2f2c1" }
